+++ src/__tests__/file.test.js
@@ -1,14 +1,11 @@
 /* eslint-env jest */
 
-const { makeExecutableSchema } = require("graphql-tools");
 const { graphql } = require("graphql");
-const { typeDefs, resolvers } = require("../schema");
+const { schema } = require("../schema");
 
 const figmaFile = "cLp23bR627jcuNSoBGkhL04E";
 
 describe("File", () => {
-    const mockSchema = makeExecutableSchema({ typeDefs, resolvers });
-
     test("can get file name", async () => {
         const query = `
             query {
@@ -18,7 +15,7 @@ describe("File", () => {
             }
         `;
 
-        await expect(graphql(mockSchema, query)).resolves.toEqual({
+        await expect(graphql(schema, query)).resolves.toEqual({
             data: { file: { name: "figma-graphql test file" } },
         });
     });