+++ src/schema.js
@@ -1,3 +1,4 @@
+const { makeExecutableSchema } = require("graphql-tools");
 const { GraphQLDateTime } = require("graphql-iso-date");
 const mergeSchema = require("./types");
 
@@ -12,6 +13,38 @@ const typeDefs = [
     type Mutation {
         noop: String
     }
+    
+    type Subscription {
+        noop: String
+    }
+    
+    enum BlendMode {
+        PASS_THROUGH
+        NORMAL
+        DARKEN
+        MULTIPLY
+        LINEAR_BURN
+        COLOR_BURN
+        LIGHTEN
+        SCREEN
+        LINEAR_DODGE
+        COLOR_DODGE
+        OVERLAY
+        SOFT_LIGHT
+        HARD_LIGHT
+        DIFFERENCE
+        EXCLUSION
+        HUE
+        SATURATION
+        COLOR
+        LUMINOSITY
+    }
+    
+    enum StrokeAlign {
+        INSIDE
+        OUTSIDE
+        CENTER
+    }
 `,
 ];
 
@@ -23,5 +56,6 @@ const resolvers = {
 };
 
 module.exports = {
-    ...mergeSchema({ typeDefs, resolvers }),
+    schema: makeExecutableSchema(mergeSchema({ typeDefs, resolvers })),
+    context: () => ({}),
 };