+++ src/types/page.js
@@ -1,23 +1,23 @@
+const { gql } = require("apollo-server-express");
 const { getChildren } = require("../utils");
 
-exports.type = `
+exports.type = gql`
     # A page inside a file
     type Page {
-        # id of the page
-        id: String!
+        id: ID!
         # name of the page
         name: String!
         # the type of the node
         type: String!
         # BG of the page
-        backgroundColor: Color,
+        backgroundColor: Color
         # A node of fixed size containing other nodes
-        frames: [Frame!]
+        frames(name: String): [Frame!]
     }
 `;
 
 exports.resolvers = {
     Page: {
-        frames: root => getChildren(root),
+        frames: (root, { name }) => getChildren(root, "children", name ? { name } : {}),
     },
 };