+++ src/types/rectangle.js
@@ -1,0 +1,26 @@
+const { gql } = require("apollo-server-express");
+const { resolvers } = require("./vector");
+
+exports.type = gql`
+    # A rectangle node
+    type Rectangle implements Node {
+        id: ID!
+        name: String!
+        visible: Boolean!
+        type: NodeType!
+        blendMode: BlendMode!
+        backgroundColor: Color!
+        fill: Color
+        position: Position
+        size: Size
+        image(params: ImageParams): String
+        cornerRadius: Int
+        rectangleCornerRadii: [Int]
+    }
+`;
+
+exports.resolvers = {
+    Rectangle: {
+        ...resolvers.Vector,
+    },
+};