+++ src/types/text.js
@@ -1,0 +1,26 @@
+const { gql } = require("apollo-server-express");
+const { resolvers } = require("./vector");
+
+exports.type = gql`
+    # A text node
+    type Text implements Node {
+        id: ID!
+        name: String!
+        visible: Boolean!
+        type: NodeType!
+        blendMode: BlendMode!
+        backgroundColor: Color!
+        fill: Color
+        position: Position
+        size: Size
+        image(params: ImageParams): String
+        characters: String
+        style: Style
+    }
+`;
+
+exports.resolvers = {
+    Text: {
+        ...resolvers.Vector,
+    },
+};