+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # [1.4.0](https://github.com/braposo/figma-graphql/compare/v1.3.0...v1.4.0) (2019-04-28)
 
-
 ### Features
 
-* updates graphql version to trigger release ([dd58628](https://github.com/braposo/figma-graphql/commit/dd58628))
+-   updates graphql version to trigger release ([dd58628](https://github.com/braposo/figma-graphql/commit/dd58628))